Tweaked some stuff, but still non-deterministic and the last tensor commutator hasnt been implemented
Former-commit-id: c13849511ebdcced1b00e76db517be647b0665ad

+++ src/HartreeFock.cc
@@ -387,7 +387,7 @@ void HartreeFock::ReorderCoefficients()
    arma::vec e_tmp = energies;
    for (int i=0;i<norbits;i++)
    {
-      float fmax = 0.0;
+      double fmax = 0.0;
       int kmax;
       int sign = 1;
       for (int k=0;k<norbits;k++)
